+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- "Convert If/Else to Ternary" now handles more complex assignments.
+
 ## [4.5.0] - 2020-07-10 - We all extract a yellow Substring ⚓
 
 ### Added